Separate CLI policy config boundaries
